+++ hazelcast/src/main/java/com/hazelcast/instance/DefaultNodeInitializer.java
@@ -19,12 +19,11 @@ package com.hazelcast.instance;
 import com.hazelcast.logging.ILogger;
 import com.hazelcast.nio.IOService;
 import com.hazelcast.nio.MemberSocketInterceptor;
-import com.hazelcast.nio.Packet;
-import com.hazelcast.nio.serialization.PortableContext;
-import com.hazelcast.nio.tcp.DefaultSocketChannelWrapper;
+import com.hazelcast.nio.tcp.DefaultPacketReader;
+import com.hazelcast.nio.tcp.DefaultPacketWriter;
+import com.hazelcast.nio.tcp.DefaultSocketChannelWrapperFactory;
 import com.hazelcast.nio.tcp.PacketReader;
 import com.hazelcast.nio.tcp.PacketWriter;
-import com.hazelcast.nio.tcp.SocketChannelWrapper;
 import com.hazelcast.nio.tcp.SocketChannelWrapperFactory;
 import com.hazelcast.nio.tcp.TcpIpConnection;
 import com.hazelcast.security.SecurityContext;
@@ -32,9 +31,6 @@ import com.hazelcast.storage.DataRef;
 import com.hazelcast.storage.Storage;
 import com.hazelcast.wan.WanReplicationService;
 import com.hazelcast.wan.impl.WanReplicationServiceImpl;
-
-import java.nio.ByteBuffer;
-import java.nio.channels.SocketChannel;
 
 public class DefaultNodeInitializer implements NodeInitializer {
 
@@ -95,18 +91,6 @@ public class DefaultNodeInitializer implements NodeInitializer {
         return new DefaultSocketChannelWrapperFactory();
     }
 
-    public static class DefaultSocketChannelWrapperFactory implements SocketChannelWrapperFactory {
-        @Override
-        public SocketChannelWrapper wrapSocketChannel(SocketChannel socketChannel, boolean client) throws Exception {
-            return new DefaultSocketChannelWrapper(socketChannel);
-        }
-
-        @Override
-        public boolean isSSlEnabled() {
-            return false;
-        }
-    }
-
     @Override
     public PacketReader createPacketReader(TcpIpConnection connection, IOService ioService) {
         return new DefaultPacketReader(connection, ioService);
@@ -115,49 +99,6 @@ public class DefaultNodeInitializer implements NodeInitializer {
     @Override
     public PacketWriter createPacketWriter(final TcpIpConnection connection, final IOService ioService) {
         return new DefaultPacketWriter();
-    }
-
-    public class DefaultPacketReader implements PacketReader {
-
-        protected final TcpIpConnection connection;
-
-        protected final IOService ioService;
-
-        protected Packet packet;
-
-        protected DefaultPacketReader(TcpIpConnection connection, IOService ioService) {
-            this.connection = connection;
-            this.ioService = ioService;
-        }
-
-        @Override
-        public void readPacket(ByteBuffer inBuffer) throws Exception {
-            while (inBuffer.hasRemaining()) {
-                if (packet == null) {
-                    packet = obtainPacket();
-                }
-                boolean complete = packet.readFrom(inBuffer);
-                if (complete) {
-                    packet.setConn(connection);
-                    ioService.handleMemberPacket(packet);
-                    packet = null;
-                } else {
-                    break;
-                }
-            }
-        }
-
-        protected Packet obtainPacket() {
-            PortableContext portableContext = node.getSerializationService().getPortableContext();
-            return new Packet(portableContext);
-        }
-    }
-
-    public static class DefaultPacketWriter implements PacketWriter {
-        @Override
-        public boolean writePacket(Packet packet, ByteBuffer socketBB) {
-            return packet.writeTo(socketBB);
-        }
     }
 
     @Override
