+++ hazelcast/src/main/java/com/hazelcast/nio/tcp/DefaultSocketChannelWrapperFactory.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright (c) 2008-2014, Hazelcast, Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.hazelcast.nio.tcp;
+
+import java.nio.channels.SocketChannel;
+
+public class DefaultSocketChannelWrapperFactory implements SocketChannelWrapperFactory {
+    @Override
+    public SocketChannelWrapper wrapSocketChannel(SocketChannel socketChannel, boolean client) throws Exception {
+        return new DefaultSocketChannelWrapper(socketChannel);
+    }
+
+    @Override
+    public boolean isSSlEnabled() {
+        return false;
+    }
+}
